chore(574): Added Edge Types
