+++ GUI/src/components/Flow/EdgeTypes/PlaceholderEdge.tsx
@@ -1,0 +1,24 @@
+import { getBezierPath, EdgeProps, BaseEdge } from "@xyflow/react";
+
+export default function PlaceholderEdge({
+  id,
+  sourceX,
+  sourceY,
+  targetX,
+  targetY,
+  sourcePosition,
+  targetPosition,
+  style,
+  markerEnd,
+}: EdgeProps) {
+  const [edgePath] = getBezierPath({
+    sourceX,
+    sourceY,
+    sourcePosition,
+    targetX,
+    targetY,
+    targetPosition,
+  });
+
+  return <BaseEdge id={id} style={style} path={edgePath} markerEnd={markerEnd} />;
+}
