+++ GUI/src/components/Flow/EdgeTypes/index.ts
@@ -1,0 +1,9 @@
+import CustomEdge from "./CustomEdge";
+import PlaceholderEdge from "./PlaceholderEdge";
+
+export const edgeTypes = {
+  placeholder: PlaceholderEdge,
+  step: CustomEdge,
+};
+
+export default edgeTypes;
