+++ GUI/src/components/Flow/NodeTypes/Node.scss
