+++ GUI/src/components/Flow/NodeTypes/StartNode.tsx
