+++ GUI/src/types/service-flow.ts
@@ -1,34 +1,17 @@
 import { MultiChoiceQuestion } from "./multi-choice-question";
 import { StepType } from "./step-type.enum";
-import { Edge, MarkerType, Node } from "@xyflow/react";
+import { Edge, Node } from "@xyflow/react";
 
 export const GRID_UNIT = 16;
 export const EDGE_LENGTH = 5 * GRID_UNIT;
 
-export const initialPlaceholder = {
-  id: "2",
-  type: "placeholder",
-  position: {
-    x: 11.12 * GRID_UNIT,
-    y: 9 * GRID_UNIT,
-  },
-  data: {
-    type: "placeholder",
-  },
-  className: "placeholder",
-  selectable: false,
-  draggable: false,
-};
-
 export const initialEdges: Edge[] = [
   {
-    type: "smoothstep",
+    type: "step",
     id: "edge-1-2",
     source: "1",
     target: "2",
-    markerEnd: {
-      type: MarkerType.ArrowClosed,
-    },
+    animated: true,
   },
 ];
 
@@ -54,15 +37,30 @@ export type NodeDataProps = {
 export const initialNodes: Node[] = [
   {
     id: "1",
-    type: "startNode",
+    type: "start",
     position: {
       x: 13.5 * GRID_UNIT,
       y: GRID_UNIT,
     },
-    data: {},
+    data: {
+      type: "start"
+    },
     className: "start",
     selectable: false,
     draggable: false,
   },
-  initialPlaceholder,
+  {
+    id: "2",
+    type: "ghost",
+    position: {
+      x: 3 * GRID_UNIT,
+      y: 11 * GRID_UNIT,
+    },
+    data: {
+      type: "ghost",
+    },
+    className: "ghost",
+    selectable: false,
+    draggable: false,
+  },
 ];
